+++ packages/moonstone/Dropdown/Dropdown.js
@@ -300,7 +300,7 @@ const DropdownBase = kind({
 		// `ui/Group`/`ui/Repeater` will throw an error if empty so we disable the Dropdown and
 		// prevent Dropdown to open if there are no children.
 		const hasChildren = children.length > 0;
-		const openDropdown = hasChildren ? open : false;
+		const openDropdown = hasChildren && !disabled && open;
 		delete rest.width;
 
 		return (